Add inverse text normalization

+++ sub_gen.py
@@ -1,16 +1,24 @@
 import os
 import subprocess
+import tqdm
 
 from asr.audio import AudioFile
 from utils import extract_audio, convert_audio, write_to_file
-import tqdm
 
 VIDEO_EXT = ['mp4', 'ogg', 'm4v', 'm4a', 'webm', 'flv', 'amv', 'avi']
 AUDIO_EXT = ['mp3', 'flac', 'wav', 'aac', 'm4a', 'weba', 'sdt']
 
 
 class SubGenerator:
-    def __init__(self, file_path, asr_model, split_duration=5000, max_words=12, sub_format=['srt'], output_directory="./temp"):
+    def __init__(self, 
+        file_path, 
+        asr_model,
+        normalizer,
+        src_lang='vi', 
+        split_duration=5000, 
+        max_words=12, 
+        sub_format=['srt'], 
+        output_directory="./temp"):
         super(SubGenerator, self).__init__()
 
         if not os.path.exists(file_path):
@@ -18,10 +26,13 @@ class SubGenerator:
 
         self.file_path = file_path
         self.file_name, self.file_ext = os.path.split(file_path)[-1].split(".")
-        self.temp_path = os.path.join("temp", self.file_name + ".wav")
-        self.model = asr_model
-        self.split_duration = split_duration
         self.max_words = max_words
+        self.temp_path = os.path.join("temp", self.file_name + ".wav")
+        self.split_duration = split_duration
+        self.src_lang = src_lang
+
+        self.model = asr_model
+        self.itn = normalizer
 
         if not os.path.exists("temp"):
             os.makedirs("temp")
@@ -38,6 +49,12 @@ class SubGenerator:
         self.sub_format = sub_format
         self.output_directory = output_directory
         self.output_file_handle_dict = {}
+    
+    def post_process(self, tokens):
+        final_tokens = self.itn.inverse_normalize_with_metadata(tokens, verbose=False)
+        final_transcript = " ".join([token['text'] for token in final_tokens])
+
+        return final_transcript, final_tokens
 
     def create_sub(self):
         self.audio_file = AudioFile(self.temp_path)
@@ -50,46 +67,52 @@ class SubGenerator:
             # For VTT format, write header
             if format == "vtt":
                 self.output_file_handle_dict[format].write("WEBVTT\n")
-                self.output_file_handle_dict[format].write(
-                    "Kind: captions\n\n")
+                self.output_file_handle_dict[format].write("Kind: captions\n\n")
         progress_bar = tqdm.tqdm(
             total=int(self.audio_file.audio_length * 1000))
         line_count = 1
         last = 0
         for start, end, audio in self.audio_file.split():
             transcript, tokens, score = self.model.transcribe_with_metadata(audio, start)[0]
-
-            if end - start > self.split_duration:
-                infer_text = ""
-                num_inferred = 0
-                prev_end = start
-
-                for token in tokens:
-                    infer_text += token['text'] + " "
-                    num_inferred += 1
-                    if num_inferred > self.max_words or token['end'] - prev_end > self.split_duration:
-                        write_to_file(self.output_file_handle_dict, infer_text,
-                                      line_count, (prev_end / 1000, token['start'] / 1000))
-                        infer_text = ""
-                        num_inferred = 0
-                        prev_end = token['end']
-                        line_count += 1
-
-                if infer_text:
-                    write_to_file(self.output_file_handle_dict, infer_text,
-                                  line_count, (prev_end / 1000, token['start'] / 1000))
-                    line_count += 1
-            else:
-                write_to_file(self.output_file_handle_dict, transcript,
-                              line_count, (start / 1000, end / 1000))
-                line_count += 1
+            transcript, tokens = self.post_process(tokens)
+            line_count = self.write_sub(transcript, tokens, start, end, line_count)
 
             progress_bar.update(int(end - last))
             last = end
 
         self.audio_file.close()
         self.close_file()
-        os.remove(self.temp_path)
+
+        if os.path.exists(self.temp_path):
+            os.remove(self.temp_path)
+    
+    def write_sub(self, transcript, tokens, start, end, line_count):
+        if end - start > self.split_duration:
+            infer_text = ""
+            num_inferred = 0
+            prev_end = start
+
+            for token in tokens:
+                infer_text += token['text'] + " "
+                num_inferred += 1
+                if num_inferred > self.max_words or token['end'] - prev_end > self.split_duration:
+                    write_to_file(self.output_file_handle_dict, infer_text,
+                                    line_count, (prev_end / 1000, token['start'] / 1000))
+                    infer_text = ""
+                    num_inferred = 0
+                    prev_end = token['end']
+                    line_count += 1
+
+            if infer_text:
+                write_to_file(self.output_file_handle_dict, infer_text,
+                                line_count, (prev_end / 1000, token['start'] / 1000))
+                line_count += 1
+        else:
+            write_to_file(self.output_file_handle_dict, transcript,
+                            line_count, (start / 1000, end / 1000))
+            line_count += 1
+        
+        return line_count
 
     def sync_sub(self):
         if "srt" not in self.sub_format:
